fixes for account microservice, improve changes

+++ diamonbank-accounts/src/main/resources/schema.sql
@@ -2,7 +2,7 @@ DROP TABLE IF EXISTS customers;
 DROP TABLE IF EXISTS accounts;
 
 CREATE TABLE customers (
-    customer_id int AUTO_INCREMENT PRIMARY KEY,
+    customer_id numeric AUTO_INCREMENT PRIMARY KEY,
     name varchar(100) NOT NULL,
     email varchar(100) NOT NULL,
     mobile_number varchar(100) NOT NULL,
@@ -11,7 +11,7 @@ CREATE TABLE customers (
 
 
 CREATE TABLE accounts (
-    account_id int AUTO_INCREMENT PRIMARY KEY,
+    account_id numeric AUTO_INCREMENT PRIMARY KEY,
     account_number varchar(20) NOT NULL,
     account_type varchar(100) NOT NULL,
     branch_address varchar(200) NOT NULL,
